requirements.txt
contourpy==1.3.2
coverage==7.9.1
cycler==0.12.1
fonttools==4.58.4
iniconfig==2.1.0
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.3.1
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
six==1.17.0
